rename pypyr.context to pypyr.parser
