+++ tests/unit/pypyr/context/json_test.py
@@ -1,5 +1,5 @@
 """json.py unit tests."""
-import pypyr.context.json
+import pypyr.parser.json
 import json
 import pytest
 
@@ -7,4 +7,4 @@ import pytest
 def test_json_cant_parse_from_arbitrary_string():
     """Comma delimited input string should fail."""
     with pytest.raises(json.JSONDecodeError):
-        pypyr.context.json.get_parsed_context('value 1,value 2, value3')
+        pypyr.parser.json.get_parsed_context('value 1,value 2, value3')
