+++ tests/unit/pypyr/context/keyvaluepairs_test.py
@@ -1,13 +1,13 @@
 """keyvaluepairs.py unit tests."""
-import pypyr.context.keyvaluepairs
+import pypyr.parser.keyvaluepairs
 import pytest
 
 
 def test_comma_string_parses_to_dict():
     """Comma delimited input kvp string should return dictionary."""
-    out = pypyr.context.keyvaluepairs.get_parsed_context('key1=value1'
-                                                         ',key2=value2'
-                                                         ',key3=value3')
+    out = pypyr.parser.keyvaluepairs.get_parsed_context('key1=value1'
+                                                        ',key2=value2'
+                                                        ',key3=value3')
     assert out['key1'] == 'value1', "key1 should be value1."
     assert out['key2'] == 'value2', "key2 should be value2."
     assert out['key3'] == 'value3', "key3 should be value3."
@@ -16,8 +16,8 @@ def test_comma_string_parses_to_dict():
 
 def test_no_commas_string_parses_to_single_entry():
     """No commas input kvp string should return dictionary with 1 item."""
-    out = pypyr.context.keyvaluepairs.get_parsed_context('key 1=value 2 '
-                                                         'value3')
+    out = pypyr.parser.keyvaluepairs.get_parsed_context('key 1=value 2 '
+                                                        'value3')
     assert out['key 1'] == 'value 2 value3', "key 1 isnt 'value 2 value3'."
     assert len(out) == 1, "1 item expected"
 
@@ -25,11 +25,11 @@ def test_no_commas_string_parses_to_single_entry():
 def test_no_equals_string_parses_to_single_entry():
     """No equals input kvp string fails with ValueError."""
     with pytest.raises(ValueError):
-        pypyr.context.keyvaluepairs.get_parsed_context(
+        pypyr.parser.keyvaluepairs.get_parsed_context(
             'key1value2,value3')
 
 
 def test_empty_string_throw():
     """Empty input string should throw assert error."""
     with pytest.raises(AssertionError):
-        pypyr.context.keyvaluepairs.get_parsed_context(None)
+        pypyr.parser.keyvaluepairs.get_parsed_context(None)
